+++ lib/Cargo.toml
@@ -42,7 +42,7 @@ hashlink = { version = "0.10.0", features = ["serde_impl"] }
 heed = "0.21.0"
 hex = { workspace = true }
 jiff = { version = "0.2.4", default-features = false }
-jsonrpsee = { workspace = true }
+jsonrpsee = { workspace = true, features = ["client"] }
 miette = { workspace = true, features = ["derive"] }
 nom = "7.1.3"
 nonempty = "0.11.0"
