+++ lib/server/wallet/grpc.rs
@@ -1,32 +1,18 @@
 use std::{borrow::Cow, collections::HashMap, str::FromStr, sync::Arc};
 
 use bdk_wallet::bip39::Mnemonic;
-use bitcoin::{
-    hashes::{hmac, ripemd160, sha256, sha512, Hash, HashEngine},
-    key::Secp256k1,
-    Address, Amount, BlockHash, OutPoint, Transaction,
-};
-use fallible_iterator::{FallibleIterator as _, IteratorExt};
+use bitcoin::{hashes::Hash as _, Address, Amount, BlockHash, Transaction};
 use futures::{
     stream::{BoxStream, FusedStream},
     StreamExt as _,
 };
 use thiserror::Error;
 
-// Re-export this with names that work better externally.
-pub use crate::server_validator::Server as Validator;
 use crate::{
     convert,
     errors::ErrorChain,
-    messages::parse_op_drivechain,
     proto::{
-        common::{ConsensusHex, Hex, ReverseHex},
-        crypto::{
-            crypto_service_server::CryptoService, HmacSha512Request, HmacSha512Response,
-            Ripemd160Request, Ripemd160Response, Secp256k1SecretKeyToPublicKeyRequest,
-            Secp256k1SecretKeyToPublicKeyResponse, Secp256k1SignRequest, Secp256k1SignResponse,
-            Secp256k1VerifyRequest, Secp256k1VerifyResponse,
-        },
+        common::ReverseHex,
         mainchain::{
             create_sidechain_proposal_response, get_info_response,
             list_sidechain_deposit_transactions_response::SidechainDepositTransaction,
@@ -45,28 +31,10 @@ use crate::{
         },
         StatusBuilder, ToStatus,
     },
+    server::{invalid_field_value, missing_field},
     types::{BlindedM6, Event, SidechainNumber},
     wallet::{error::WalletInitialization, CreateTransactionParams},
 };
-
-pub(crate) fn invalid_field_value<Message, Error>(
-    field_name: &str,
-    value: &str,
-    source: Error,
-) -> tonic::Status
-where
-    Message: prost::Name,
-    Error: std::error::Error + Send + Sync + 'static,
-{
-    crate::proto::Error::invalid_field_value::<Message, _>(field_name, value, source).into()
-}
-
-pub(crate) fn missing_field<Message>(field_name: &str) -> tonic::Status
-where
-    Message: prost::Name,
-{
-    crate::proto::Error::missing_field::<Message>(field_name).into()
-}
 
 /// Stream (non-)confirmations for a sidechain proposal
 fn stream_proposal_confirmations(
@@ -628,72 +596,17 @@ impl WalletService for crate::wallet::Wallet {
     ) -> Result<tonic::Response<ListSidechainDepositTransactionsResponse>, tonic::Status> {
         let ListSidechainDepositTransactionsRequest {} = request.into_inner();
         let transactions = self
-            .list_wallet_transactions()
+            .list_sidechain_deposit_transactions()
             .await
             .map_err(|err| err.builder().to_status())?
             .into_iter()
-            .map(Ok::<_, tonic::Status>)
-            .transpose_into_fallible()
-            .filter_map(|bdk_wallet_tx| {
-                let Some(treasury_output) = bdk_wallet_tx.tx.output.first() else {
-                    return Ok(None);
-                };
-                let Ok((_, sidechain_number)) =
-                    parse_op_drivechain(&treasury_output.script_pubkey.to_bytes())
-                else {
-                    return Ok(None);
-                };
-                let treasury_outpoint = OutPoint {
-                    txid: bdk_wallet_tx.txid,
-                    vout: 0,
-                };
-                let spent_ctip = match self
-                    .validator()
-                    .try_get_ctip_value_seq(&treasury_outpoint)
-                    .map_err(|err| err.builder().to_status())?
-                {
-                    Some((_, _, seq)) => {
-                        let spent_treasury_utxo = self
-                            .validator()
-                            .get_treasury_utxo(sidechain_number, seq - 1)
-                            .map_err(|err| err.builder().to_status())?;
-                        Some(crate::types::Ctip {
-                            outpoint: spent_treasury_utxo.outpoint,
-                            value: spent_treasury_utxo.total_value,
-                        })
-                    }
-                    None => {
-                        // May be unconfirmed
-                        // check if current ctip in inputs
-                        match self
-                            .validator()
-                            .try_get_ctip(sidechain_number)
-                            .map_err(|err| err.builder().to_status())?
-                        {
-                            Some(ctip) => {
-                                if bdk_wallet_tx.tx.input.iter().any(|txin: &bitcoin::TxIn| {
-                                    txin.previous_output == ctip.outpoint
-                                }) {
-                                    Some(ctip)
-                                } else {
-                                    return Ok(None);
-                                }
-                            }
-                            None => None,
-                        }
-                    }
-                };
-                if let Some(spent_ctip) = spent_ctip {
-                    if spent_ctip.value > treasury_output.value {
-                        return Ok(None);
-                    }
-                }
-                Ok(Some(SidechainDepositTransaction {
+            .map(
+                |(sidechain_number, bdk_wallet_tx)| SidechainDepositTransaction {
                     sidechain_number: Some(sidechain_number.0.into()),
                     tx: Some(WalletTransaction::from(&bdk_wallet_tx)),
-                }))
-            })
-            .collect()?;
+                },
+            )
+            .collect();
         let response = ListSidechainDepositTransactionsResponse { transactions };
         Ok(tonic::Response::new(response))
     }
@@ -900,146 +813,4 @@ enum Error {
 
     #[error("address must be non-empty")]
     AddressMustBeNonEmpty,
-}
-
-#[derive(Debug, Default)]
-pub struct CryptoServiceServer;
-
-const SECP256K1_SECRET_KEY_LENGTH: usize = 32;
-const SECP256K1_PUBLIC_KEY_LENGTH: usize = 33;
-const SECP256K1_SIGNATURE_COMPACT_SERIALIZATION_LENGTH: usize = 64;
-
-#[tonic::async_trait]
-impl CryptoService for CryptoServiceServer {
-    async fn ripemd160(
-        &self,
-        request: tonic::Request<Ripemd160Request>,
-    ) -> std::result::Result<tonic::Response<Ripemd160Response>, tonic::Status> {
-        let Ripemd160Request { msg } = request.into_inner();
-        let msg: Vec<u8> = msg
-            .ok_or_else(|| missing_field::<Ripemd160Request>("msg"))?
-            .decode_tonic::<Ripemd160Request, _>("msg")?;
-        let digest = ripemd160::Hash::hash(&msg);
-        let response = Ripemd160Response {
-            digest: Some(Hex::encode(&digest.as_byte_array())),
-        };
-        Ok(tonic::Response::new(response))
-    }
-
-    async fn hmac_sha512(
-        &self,
-        request: tonic::Request<HmacSha512Request>,
-    ) -> std::result::Result<tonic::Response<HmacSha512Response>, tonic::Status> {
-        let HmacSha512Request { key, msg } = request.into_inner();
-        let key: Vec<u8> = key
-            .ok_or_else(|| missing_field::<HmacSha512Request>("key"))?
-            .decode_tonic::<HmacSha512Request, _>("key")?;
-        let msg: Vec<u8> = msg
-            .ok_or_else(|| missing_field::<HmacSha512Request>("msg"))?
-            .decode_tonic::<HmacSha512Request, _>("msg")?;
-        let mut engine = hmac::HmacEngine::<sha512::Hash>::new(&key);
-        engine.input(&msg);
-        let hmac = hmac::Hmac::<sha512::Hash>::from_engine(engine);
-        let response = HmacSha512Response {
-            hmac: Some(Hex::encode(&hmac.as_byte_array())),
-        };
-        Ok(tonic::Response::new(response))
-    }
-
-    async fn secp256k1_secret_key_to_public_key(
-        &self,
-        request: tonic::Request<Secp256k1SecretKeyToPublicKeyRequest>,
-    ) -> std::result::Result<tonic::Response<Secp256k1SecretKeyToPublicKeyResponse>, tonic::Status>
-    {
-        let Secp256k1SecretKeyToPublicKeyRequest { secret_key } = request.into_inner();
-        let secret_key: [u8; SECP256K1_SECRET_KEY_LENGTH] = secret_key
-            .ok_or_else(|| missing_field::<Secp256k1SecretKeyToPublicKeyRequest>("secret_key"))?
-            .decode_tonic::<Secp256k1SecretKeyToPublicKeyRequest, _>("secret_key")?;
-        let secret_key = bitcoin::secp256k1::SecretKey::from_slice(&secret_key).map_err(|err| {
-            invalid_field_value::<Secp256k1SecretKeyToPublicKeyRequest, _>("secret_key", "", err)
-        })?;
-        let secp = Secp256k1::new();
-        let public_key = bitcoin::secp256k1::PublicKey::from_secret_key(&secp, &secret_key);
-        let public_key: [u8; SECP256K1_PUBLIC_KEY_LENGTH] = public_key.serialize();
-        let response = Secp256k1SecretKeyToPublicKeyResponse {
-            public_key: Some(ConsensusHex::encode(&public_key)),
-        };
-        Ok(tonic::Response::new(response))
-    }
-
-    async fn secp256k1_sign(
-        &self,
-        request: tonic::Request<Secp256k1SignRequest>,
-    ) -> std::result::Result<tonic::Response<Secp256k1SignResponse>, tonic::Status> {
-        let Secp256k1SignRequest {
-            message,
-            secret_key,
-        } = request.into_inner();
-        let message: Vec<u8> = message
-            .ok_or_else(|| missing_field::<Secp256k1SignRequest>("message"))?
-            .decode_tonic::<Secp256k1SignRequest, _>("message")?;
-        let digest = sha256::Hash::hash(&message).to_byte_array();
-        let message = bitcoin::secp256k1::Message::from_digest(digest);
-        let secret_key: [u8; SECP256K1_SECRET_KEY_LENGTH] = secret_key
-            .ok_or_else(|| missing_field::<Secp256k1SignRequest>("secret_key"))?
-            .decode_tonic::<Secp256k1SignRequest, _>("secret_key")?;
-        let secret_key = bitcoin::secp256k1::SecretKey::from_slice(&secret_key)
-            .map_err(|err| invalid_field_value::<Secp256k1SignRequest, _>("secret_key", "", err))?;
-        let secp = Secp256k1::new();
-        let signature = secp.sign_ecdsa(&message, &secret_key);
-        let signature = signature.serialize_compact();
-        let response = Secp256k1SignResponse {
-            signature: Some(Hex::encode(&signature)),
-        };
-        Ok(tonic::Response::new(response))
-    }
-
-    async fn secp256k1_verify(
-        &self,
-        request: tonic::Request<Secp256k1VerifyRequest>,
-    ) -> std::result::Result<tonic::Response<Secp256k1VerifyResponse>, tonic::Status> {
-        let Secp256k1VerifyRequest {
-            message,
-            signature,
-            public_key,
-        } = request.into_inner();
-        let message: Vec<u8> = message
-            .ok_or_else(|| missing_field::<Secp256k1VerifyRequest>("message"))?
-            .decode_tonic::<Secp256k1VerifyRequest, _>("message")?;
-        let digest = sha256::Hash::hash(&message).to_byte_array();
-        let message = bitcoin::secp256k1::Message::from_digest(digest);
-        let signature: Vec<u8> = signature
-            .ok_or_else(|| missing_field::<Secp256k1VerifyRequest>("signature"))?
-            .decode_tonic::<Secp256k1VerifyRequest, _>("signature")?;
-        let signature_len = signature.len();
-        let signature: [u8; SECP256K1_SIGNATURE_COMPACT_SERIALIZATION_LENGTH] =
-            signature.try_into().map_err(|_err| {
-                tonic::Status::new(
-                    tonic::Code::InvalidArgument,
-                    format!("invalid signature length {signature_len}, must be {SECP256K1_SIGNATURE_COMPACT_SERIALIZATION_LENGTH}"),
-                )
-            })?;
-        let signature =
-            bitcoin::secp256k1::ecdsa::Signature::from_compact(&signature).map_err(|err| {
-                invalid_field_value::<Secp256k1VerifyRequest, _>(
-                    "signature",
-                    &hex::encode(signature),
-                    err,
-                )
-            })?;
-        let public_key: [u8; SECP256K1_PUBLIC_KEY_LENGTH] = public_key
-            .ok_or_else(|| missing_field::<Secp256k1VerifyRequest>("public_key"))?
-            .decode_tonic::<Secp256k1VerifyRequest, _>("public_key")?;
-        let public_key = bitcoin::secp256k1::PublicKey::from_slice(&public_key).map_err(|err| {
-            invalid_field_value::<Secp256k1VerifyRequest, _>(
-                "public_key",
-                &hex::encode(public_key),
-                err,
-            )
-        })?;
-        let secp = Secp256k1::new();
-        let valid = secp.verify_ecdsa(&message, &signature, &public_key).is_ok();
-        let response = Secp256k1VerifyResponse { valid };
-        Ok(tonic::Response::new(response))
-    }
 }
